Moved CurrentPilotLogic to PilotController and Moved Pilot to Request Context
